feat: Implement Linux build and support via Nix
This commit introduces basic support for the Linux platform, using the Nix package manager and Flakes to create reproducible builds and a consistent development environment.
The primary goal of this change is to enable the application to be built, packaged, and run on Linux.
Currently, an overlay of mine was used to update python dependencies

**Implementation via Nix:**
- `flake.nix` exposes both package and homeModule outputs
- A `devShell` is now provided, accessible via `nix develop`, to ensure a consistent environment
- The Home Manager module allows for installation and configuration via nixos

**Code Changes for Linux Compatibility:**
To support running from the read-only `/nix/store`, the following changes were made - only when running on linux:
- Configuration is resolved from `~/.config/adb_auto_player`
- Debug screenshots are written to `~/.local/state/adb_auto_player`
- Templates are decoupled from the Configuration and are now stored in the `$out/bin/templates` folder

**Known Issues:**
- [ERROR] Failed to create debug directory: mkdir debug: read-only file system
- [ERROR] Failed to setup log file: mkdir debug: read-only file system
- These do not seem to be breaking